frontend/requirements.txt
streamlit==1.32.0
supabase==2.3.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.1
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
requests==2.31.0